Expose visit api
Adds a method that can be used to visit data from a cluster. The visit
method returns a generator yielding each slice, which in turn returns a
generator yielding each response. Optionally run concurrently by
increasing the number of slices, and optionally supply a yql query, or
schema to filter.